requirements.txt
pytesseract==0.3.10
Pillow==9.4.0
requests
tqdm